requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
beautifulsoup4 
langchain 
langchain-groq
langchain-google-community
langchain-google-genai
fastapi
uvicorn
supabase
APScheduler